refactor(core): rename history::non_atomic → history::raw
The `non_atomic` module name was misleading — it represents the raw input
history format, not a non-atomic property. Rename to `raw` for clarity.

All internal and external import paths updated. No logic changes.
