+++ dbcop_core/src/history/mod.rs
@@ -1,2 +1,2 @@
 pub mod atomic;
-pub mod non_atomic;
+pub mod raw;
